added update also minions without targets at once

+++ spmigration/spmigration.go
@@ -190,11 +190,18 @@ func (m *Target_Minions) Get_Minions(sessionkey *auth.SumaSessionKey, groupsdata
 				// Access specific member values by name
 				minion_data.Minion_Name = valueStruct.GetMemberValue("name").(string)
 				minion_data.Minion_ID = valueStruct.GetMemberValue("id").(int)
-				all_minions = append(all_minions, minion_data)
-				Delete_Notes(sessionkey, minion_data.Minion_ID)
-				all_minions_in_group[group] = append(all_minions_in_group[group], minion_data.Minion_Name)
-				//logger.Infof("name: %s, id: %d\n", minion_data.Minion_Name, minion_data.Minion_ID)
-				//if Contains(active_minion_ids, minion_data.Minion_ID) {
+				found_existing_minion := false
+				for _, existing_minion := range all_minions {
+					if existing_minion.Minion_ID == minion_data.Minion_ID {
+						logger.Infof("Minion %s already exists in Minion_List\n", minion_data.Minion_Name)
+						found_existing_minion = true
+					}
+				}
+				if !found_existing_minion {
+					all_minions = append(all_minions, minion_data)
+					Delete_Notes(sessionkey, minion_data.Minion_ID)
+					all_minions_in_group[group] = append(all_minions_in_group[group], minion_data.Minion_Name)
+				}
 
 			}
 			// write all minions in group to a yaml file with group name
@@ -212,29 +219,39 @@ func (m *Target_Minions) Get_Minions(sessionkey *auth.SumaSessionKey, groupsdata
 		}
 
 	}
+	//logger.Infof("all_minions: %v\n", all_minions)
 	online_minions, offline_minions := m.Detect_Online_Minions(sessionkey, all_minions, groupsdata)
 
+	//logger.Infof("online_minions: %v\n", online_minions)
 	for _, minion_data := range online_minions {
 		if minion_data.Minion_ID != 0 {
-			ident, target_migration_base_channel := Find_MigrationTarget(sessionkey, minion_data.Minion_ID, groupsdata)
-			if ident != "" && target_migration_base_channel != "" {
-				valid_target_minions = append(valid_target_minions, minion_data)
-				logger.Infof("Minion %s has a valid migration target %s\n",
-					minion_data.Minion_Name, target_migration_base_channel)
+			if groupsdata.Include_Spmigration {
+				ident, target_migration_base_channel := Find_MigrationTarget(sessionkey, minion_data.Minion_ID, groupsdata)
+				if ident != "" && target_migration_base_channel != "" {
+					valid_target_minions = append(valid_target_minions, minion_data)
+					logger.Infof("Minion %s has a valid migration target %s\n",
+						minion_data.Minion_Name, target_migration_base_channel)
+				} else {
+					no_target_minions = append(no_target_minions, minion_data)
+					subject := "no valid migration target"
+					body := "minion does not have a valid migration target."
+					Add_Note(sessionkey, minion_data.Minion_ID, subject, body)
+					logger.Infof("Minion %s has not a valid migration target\n", minion_data.Minion_Name)
+				}
 			} else {
-				no_target_minions = append(no_target_minions, minion_data)
-				subject := "no valid migration target"
-				body := "minion does not have a valid migration target."
-				Add_Note(sessionkey, minion_data.Minion_ID, subject, body)
-				logger.Infof("Minion %s has not a valid migration target\n", minion_data.Minion_Name)
+				valid_target_minions = append(valid_target_minions, minion_data)
 			}
 		}
 	}
 
-	//logger.Infof("Minions in group %s: %v\n", group, all_minions_in_group[group])
-	m.Add_Online_Minions(valid_target_minions)
 	m.Add_Offline_Minions(offline_minions)
-	m.Add_No_Target_Minions(no_target_minions)
+	if groupsdata.Include_Spmigration {
+		m.Add_Online_Minions(valid_target_minions)
+		m.Add_No_Target_Minions(no_target_minions)
+	} else {
+		m.Add_No_Target_Minions(valid_target_minions)
+		logger.Infof("Adding %d online minions for only update workflow.\n", len(m.No_Targets_Minions))
+	}
 	//m.Show_Minions()
 	return nil
 }
@@ -242,12 +259,12 @@ func (m *Target_Minions) Get_Minions(sessionkey *auth.SumaSessionKey, groupsdata
 func (s *Target_Minions) Show_Minions() {
 	logger.Infof("Online Minions with targets:\n")
 	for _, minion := range s.Minion_List {
-		logger.Infof("Minion name: %s, Minion ID: %d\n", minion.Minion_Name, minion.Minion_ID)
+		logger.Infof("With Target Minion name: %s, Minion ID: %d\n", minion.Minion_Name, minion.Minion_ID)
 	}
 
 	logger.Infof("Online Minions without targets:\n")
 	for _, minion := range s.No_Targets_Minions {
-		logger.Infof("Minion name: %s, Minion ID: %d\n", minion.Minion_Name, minion.Minion_ID)
+		logger.Infof("No Target Minion name: %s, Minion ID: %d\n", minion.Minion_Name, minion.Minion_ID)
 	}
 }
 
@@ -305,12 +322,9 @@ func Orchestrate(sessionkey *auth.SumaSessionKey, groupsdata *Migration_Groups, 
 
 	target_minions.Get_Minions(sessionkey, groupsdata)
 	//logger.Infof("Minions in group: %v\n", target_minions.Minion_List)
-	logger.Infof("what is val of qualifying: %v\n", groupsdata.Qualifying_only)
-	if groupsdata.Qualifying_only {
-		logger.Infof("Qualifying only is set true so we exit here.\n")
-		return
-	}
-	target_minions.Show_Minions()
+	logger.Infof("Qualifying only value set to: %v\n", groupsdata.Qualifying_only)
+
+	//target_minions.Show_Minions()
 	target_minions.Write_Tracking_file()
 	target_minions.Salt_Refresh_Grains(sessionkey, groupsdata)
 	target_minions.Salt_No_Upgrade_Exception_Check(sessionkey, groupsdata)
@@ -318,13 +332,21 @@ func Orchestrate(sessionkey *auth.SumaSessionKey, groupsdata *Migration_Groups, 
 	target_minions.SPMigration_Group(sessionkey, groupsdata)
 
 	target_minions.Salt_Run_state_apply(sessionkey, groupsdata, "pre")
-	//target_minions.Show_Minions()
+	target_minions.Show_Minions()
 	target_minions.Write_Tracking_file()
 	emails.SPmigration_Tracking_File = target_minions.Tracking_file_name
 	emails.Template_dir = email_template_dir
-	target_minions.Assign_Channels(sessionkey, groupsdata)
+
 	emails.Send_SPmigration_Email()
-	target_minions.Check_Assigne_Channels_Jobs(sessionkey, health) // deadline 15min
+	if groupsdata.Qualifying_only {
+		logger.Infof("Qualifying only is set true so we exit here.\n")
+		return
+	}
+	if groupsdata.Include_Spmigration {
+		target_minions.Assign_Channels(sessionkey, groupsdata)
+		target_minions.Check_Assigne_Channels_Jobs(sessionkey, health) // deadline 10min
+	}
+
 	//target_minions.Schedule_Pkg_refresh(sessionkey)        // pkg refresh
 	//target_minions.Check_Pkg_Refresh_Jobs(sessionkey)      // deadline 15min
 	JobID_Pkg_Update := target_minions.Schedule_Package_Updates(sessionkey)
@@ -333,16 +355,23 @@ func Orchestrate(sessionkey *auth.SumaSessionKey, groupsdata *Migration_Groups, 
 	target_minions.Check_Reboot_Jobs(sessionkey, health)
 	target_minions.Schedule_Pkg_refresh(sessionkey)           // pkg refresh
 	target_minions.Check_Pkg_Refresh_Jobs(sessionkey, health) // deadline 15min
-	target_minions.ListMigrationTarget(sessionkey, groupsdata)
-	target_minions.Schedule_Migration(sessionkey, groupsdata, true)
-	target_minions.Check_SP_Migration(sessionkey, true, health)
-	target_minions.Schedule_Migration(sessionkey, groupsdata, false)
-	target_minions.Check_SP_Migration(sessionkey, false, health)
-	target_minions.Salt_Set_Patch_Level(sessionkey, groupsdata)
-	target_minions.Salt_Refresh_Grains(sessionkey, groupsdata)
-	target_minions.Schedule_Reboot(sessionkey)
-	target_minions.Check_Reboot_Jobs(sessionkey, health)
-	target_minions.Analyze_Pending_SPMigration(sessionkey, groupsdata, email_template_dir, health)
+
+	if groupsdata.Include_Spmigration {
+		target_minions.ListMigrationTarget(sessionkey, groupsdata)
+		target_minions.Schedule_Migration(sessionkey, groupsdata, true)
+		target_minions.Check_SP_Migration(sessionkey, true, health)
+		target_minions.Schedule_Migration(sessionkey, groupsdata, false)
+		target_minions.Check_SP_Migration(sessionkey, false, health)
+		target_minions.Salt_Set_Patch_Level(sessionkey, groupsdata)
+		target_minions.Salt_Refresh_Grains(sessionkey, groupsdata)
+		target_minions.Schedule_Reboot(sessionkey)
+		target_minions.Check_Reboot_Jobs(sessionkey, health)
+		target_minions.Analyze_Pending_SPMigration(sessionkey, groupsdata, email_template_dir, health)
+	} else {
+		target_minions.Salt_Set_Patch_Level(sessionkey, groupsdata)
+		target_minions.Salt_Refresh_Grains(sessionkey, groupsdata)
+	}
+
 	target_minions.Salt_CSV_Report(sessionkey, groupsdata)
 	target_minions.Write_Tracking_file()
 	target_minions.Salt_Run_state_apply(sessionkey, groupsdata, "post")
